recipe/serializers.py
from .models import Recipe
from rest_framework import serializers


class RecipeSerializer(serializers.ModelSerializer):
    
    class Meta:
        model = Recipe
        fields = ['author','title', 'ingredients', 'tag', 'time_to_done']