pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "foodgram project"
authors = ["orcakid <orcakid97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
